Genre and RecordLabel CRUD methods created

+++ Database_Project_Group1/src/Genre.java
@@ -1,3 +1,4 @@
+import java.sql.Date;
 import java.util.ArrayList;
 
 public class Genre {
@@ -29,93 +30,105 @@ public class Genre {
     }
 
     /**
-     * selectAll method of the Genre class.
+     * selectGenre method of the Genre class.
      * This method will use the getData(String, ArrayList<String>) method of the Alinity
-     * class. The SQL statement will select Genres where the ID value input will be bound
+     * class. The SQL statement will select Genres where the genreName value input will be bound
      * from the ArrayList, and save the result.
      * Sets genreID, name based on result data.
-     * This method will specifically work only when selecting ALL the information for said genre
-     * id.
      * If the attempt of the executing the statement fails, log the error to the file.
      *
      * @throws AlinityException
      */
-    public void selectAll(ArrayList<String> stringList) throws AlinityException {
+    public void selectGenre(User user, String genreName) throws AlinityException {
         try {
-            ArrayList<ArrayList<String>> result = AlinityMain.alinityDB.getData("SELECT * FROM Genre WHERE genreId = ?" , stringList);
-            System.out.print("\nColumn headers: " + result.get(0));
-            ArrayList<String> genreData = result.get(1);
-            setGenreId(Integer.parseInt(genreData.get(0)));
-            setGenreName(genreData.get(1));
+            if(user.getRole().equals("General") || user.getRole().equals("Admin")){
+                ArrayList<String> info = new ArrayList<>();
+                info.add(genreName);
+                ArrayList<ArrayList<String>> result = AlinityMain.alinityDB.getData("SELECT * FROM Album WHERE genreName = ?" , info);
+                System.out.print("\nColumn headers: " + result.get(0));
+                ArrayList<String> genreData = result.get(1);
+                setGenreId(Integer.parseInt(genreData.get(0)));
+                setGenreName(genreData.get(1));
+                printGenre();
+            } else System.out.println("You do not have access to this function. Please contact an administrator.");
         } catch (IndexOutOfBoundsException ioobe) {
-            throw new AlinityException(ioobe, "-> Error in obtaining data (IndexOutOfBoundsException) from the database. Please check your syntax in the selectAll(ArrayList<String>) method.", "SELECT * FROM Genre WHERE genreId = ?");
+            throw new AlinityException(ioobe, "-> Error in obtaining data (IndexOutOfBoundsException) from the database. Please check your syntax in the selectGenre(ArrayList<String>) method.", "SELECT * FROM Album WHERE genreName = ?");
         } catch (NullPointerException npe) {
-            throw new AlinityException(npe, "-> Error in obtaining data (NullPointerException) from the database. Please check your syntax in the selectAll(ArrayList<String>) method.", "SELECT * FROM Genre WHERE genreId = ?");
+            throw new AlinityException(npe, "-> Error in obtaining data (NullPointerException) from the database. Please check your syntax in the selectGenre(ArrayList<String>) method.", "SELECT * FROM Album WHERE genreName = ?");
         }
     }
 
     /**
-     * updateAll method of the Genre class.
+     * updateGenre method of the Genre class.
      * A SQL statement as a String is created using the current name
      * where the genreId is based on the values which will be bound
      * from the input ArrayList of Strings.
      * Executes using the setData method.
-     * This method will specifically work only when updating ALL the information for said genre
-     * id.
      * If the attempt of the executing the statement fails, log the error to the file.
      *
      * @throws AlinityException
      */
-    public boolean updateAll(ArrayList<String> stringList) throws AlinityException {
+    public boolean updateGenre(User user, String genreName) throws AlinityException {
         try {
-            String putStmt = "UPDATE Genre SET genreName = ? WHERE genreId = ?";
-            return AlinityMain.alinityDB.setData(putStmt, stringList);
+            if(user.getRole().equals("Admin")) {
+                ArrayList<String> info = new ArrayList<>();
+                info.add(genreName);
+                info.add(String.valueOf(this.getGenreId()));
+                String putStmt = "UPDATE Genre SET genreName = ? WHERE genreId = ?";
+                return AlinityMain.alinityDB.setData(putStmt, info);
+            } else System.out.println("You do not have the correct permissions to use this function. Please contact an administrator."); return false;
         } catch (NullPointerException npe) {
-            throw new AlinityException(npe, "-> Error in manipulating data (NullPointerException) from the database. Please check your syntax in the updateAll(ArrayList<String>) method.", "UPDATE Genre SET genreName = ? WHERE genreId = ?");
+            throw new AlinityException(npe, "-> Error in manipulating data (NullPointerException) from the database. Please check your syntax in the updateGenre(ArrayList<String>) method.", "UPDATE Genre SET genreName = ? WHERE genreId = ?");
         } catch (IndexOutOfBoundsException ioobe) {
-            throw new AlinityException(ioobe, "-> Error in manipulating data (IndexOutOfBoundsException) from the database. Please check your syntax in the updateAll(ArrayList<String>) method.", "UPDATE Genre SET genreName = ? WHERE genreId = ?");
+            throw new AlinityException(ioobe, "-> Error in manipulating data (IndexOutOfBoundsException) from the database. Please check your syntax in the updateGenre(ArrayList<String>) method.", "UPDATE Genre SET genreName = ? WHERE genreId = ?");
         }
     }
 
     /**
-     * insertAll method of the Genre class.
+     * insertGenre method of the Genre class.
      * A SQL statement as a String is created, which takes in
      * values based on the values to be bound from the ArrayList of Strings.
      * Executes using the setData method.
-     * This method will specifically work only when inserting ALL the information for said genre.
      * If the attempt of the executing the statement fails, log the error to the file.
      *
      * @throws AlinityException
      */
-    public boolean insertAll(ArrayList<String> stringList) throws AlinityException {
+    public boolean insertGenre(User user, String genreName) throws AlinityException {
         try {
-            String insertStmt = "INSERT INTO Genre (genreName) VALUES (?)";
-            return AlinityMain.alinityDB.setData(insertStmt, stringList);
+            if(user.getRole().equals("Admin")) {
+                ArrayList<String> info = new ArrayList<>();
+                info.add(genreName);
+                String insertStmt = "INSERT INTO Album (genreName) VALUES (?)";
+                return AlinityMain.alinityDB.setData(insertStmt, info);
+            } else System.out.println("You do not have the correct permissions to use this function. Please contact an administrator."); return false;
         } catch (NullPointerException npe) {
-            throw new AlinityException(npe, "-> Error in manipulating data (NullPointerException) from the database. Please check your syntax in the insertAll(ArrayList<String>) method.","INSERT INTO Genre (genreName) VALUES (?)");
+            throw new AlinityException(npe, "-> Error in manipulating data (NullPointerException) from the database. Please check your syntax in the insertGenre(ArrayList<String>) method.","INSERT INTO Album (genreName) VALUES (?)");
         } catch (IndexOutOfBoundsException ioobe) {
-            throw new AlinityException(ioobe, "-> Error in manipulating data (IndexOutOfBoundsException) from the database. Please check your syntax in the insertAll(ArrayList<String>) method.","INSERT INTO Genre (genreName) VALUES (?)");
+            throw new AlinityException(ioobe, "-> Error in manipulating data (IndexOutOfBoundsException) from the database. Please check your syntax in the insertGenre(ArrayList<String>) method.","INSERT INTO Album (genreName) VALUES (?)");
         }
     }
 
     /**
-     * deleteAll method of the Genre class.
+     * deleteAlbum method of the Genre class.
      * Deletes the given album data where the genreId is bound
      * via tha value inside the ArrayList of Strings.
      * Executes using the setData method.
-     * This method will specifically work only when deleting ALL the information for said genre id
      * If the attempt of the executing the statement fails, log the error to the file.
      *
      * @throws AlinityException
      */
-    public boolean deleteAll(ArrayList<String> stringList) throws AlinityException {
+    public boolean deleteAlbum(User user) throws AlinityException {
         try {
-            String deleteStmt = "DELETE FROM Genre WHERE genreId = ?";
-            return AlinityMain.alinityDB.setData(deleteStmt, stringList);
+            if(user.getRole().equals("Admin")) {
+                ArrayList<String> info = new ArrayList<>();
+                info.add(String.valueOf(this.getGenreId()));
+                String deleteStmt = "DELETE FROM Genre WHERE genreId = ?";
+                return AlinityMain.alinityDB.setData(deleteStmt, info);
+            } else System.out.println("You do not have the correct permissions to use this function. Please contact an administrator."); return false;
         } catch (NullPointerException npe) {
-            throw new AlinityException(npe, "-> Error in manipulating data (NullPointerException) from the database. Please check your syntax in the deleteAll() method.", "DELETE FROM Genre WHERE genreId = ?");
+            throw new AlinityException(npe, "-> Error in manipulating data (NullPointerException) from the database. Please check your syntax in the deleteAlbum() method.", "DELETE FROM Genre WHERE genreId = ?");
         } catch (IndexOutOfBoundsException ioobe) {
-            throw new AlinityException(ioobe, "-> Error in manipulating data (IndexOutOfBoundsException) from the database. Please check your syntax in the deleteAll() method.", "DELETE FROM Genre WHERE genreId = ?");
+            throw new AlinityException(ioobe, "-> Error in manipulating data (IndexOutOfBoundsException) from the database. Please check your syntax in the deleteAlbum() method.", "DELETE FROM Genre WHERE genreId = ?");
         }
     }
